#300 - mu: integrate NSMaps and Namespace
