+++ src/mu/core/functions.rs
@@ -11,6 +11,7 @@ use crate::{
         frame::{Frame, MuFunction as _},
         indirect::MuFunction as _,
         mu::{Core as _, Mu},
+        namespace::{Core as _, MuFunction as _, Namespace},
         types::{MuFunction as _, Tag, Type},
     },
     system::sys::System,
@@ -20,7 +21,6 @@ use crate::{
         fixnum::{Core as _, Fixnum, MuFunction as _},
         float::{Core as _, Float, MuFunction as _},
         function::{Core as _, Function},
-        namespace::{Core as _, MuFunction as _, Namespace},
         stream::{Core as _, MuFunction as _, Stream},
         struct_::{Core as _, MuFunction as _, Struct},
         symbol::{Core as _, MuFunction as _, Symbol},
