+++ src/mu/core/mu.rs
@@ -10,7 +10,7 @@ use {
             exception::{Condition, Exception},
             frame::Frame,
             functions::{Core as _, LibMuFunction},
-            nsmap::NSMaps,
+            namespace::{Core as _, NSMaps, Namespace},
             reader::{Core as _, Reader},
             types::{Tag, Type},
         },
@@ -22,7 +22,6 @@ use {
             fixnum::{Core as _, Fixnum},
             float::{Core as _, Float},
             function::{Core as _, Function},
-            namespace::{Core as _, Namespace},
             stream::{Core as _, Stream},
             streambuilder::StreamBuilder,
             struct_::{Core as _, Struct},
