+++ src/mu/core/reader.rs
@@ -8,6 +8,7 @@ use crate::{
         direct::DirectType,
         exception::{self, Condition, Exception},
         mu::Mu,
+        namespace::{Core as _, Namespace},
         readtable::{map_char_syntax, SyntaxType},
         types::{Tag, Type},
     },
@@ -16,7 +17,6 @@ use crate::{
         cons::{Cons, Core as _},
         fixnum::Fixnum,
         float::Float,
-        namespace::{Core as _, Namespace},
         stream::{Core as _, Stream},
         streambuilder::StreamBuilder,
         struct_::{Core as _, Struct},
