+++ src/mu/types/symbol.rs
@@ -9,13 +9,12 @@ use {
             frame::Frame,
             indirect::IndirectTag,
             mu::{Core as _, Mu},
-            nsmap::NSMaps,
+            namespace::{Core as _, NSMaps, Namespace},
             readtable::{map_char_syntax, SyntaxType},
             types::{Tag, TagType, Type},
         },
         types::{
             char::Char,
-            namespace::{Core as _, Namespace},
             stream::{Core as _, Stream},
             vecimage::{TypedVec, VecType},
             vector::{Core as _, Vector},
